memory: remove template argument from FreeList
Template argument for FreeList was used only to calculate necessity of
pages squashing in the free list. This can be achieved via passing
boolean flag to initialization routine.

Activity: #30
