+++ kernel/CMakeLists.txt
@@ -86,6 +86,7 @@ set(kernel_sources ${kernel_start_file}
 	
 	buddy_allocator.hpp
 	free_list.hpp
+	free_list.cpp
 	page_allocator.hpp
 	page_allocator.cpp
 	page.cpp
