+++ kernel/free_list.hpp
@@ -3,6 +3,7 @@
 #include <bolgenos-ng/error.h>
 
 #include <bolgenos-ng/cout.hpp>
+#include <bolgenos-ng/page.hpp>
 #include <bolgenos-ng/stdtypes.hpp>
 
 #include "config.h"
@@ -18,16 +19,10 @@ namespace allocators {
 ///
 /// The structure provides functionality of free list list allocator. Such
 /// allocator keeps one-directional list of blocks of page frames.
-/// \tparam MaxOrder Maximal order of the free list in set. TODO: to remove.
-template<size_t MaxOrder>
 class FreeList {
 	struct item_type; // forward declaration.
-public:
-	struct max_order {
-		/// Maximal order of the free list in set. TODO: to remove.
-		static constexpr size_t value = MaxOrder;
-	};
 
+public:
 
 	/// Structure that holds statistics of the usage of the free list.
 	struct stats_type {
@@ -62,16 +57,11 @@ public:
 	/// the specified order.
 	///
 	/// \param order Order of free list.
-	/// \return boolean status of initalization. true if success; false
+	/// \param disable_squashing Boolean flag shows whether should
+	/// page squashing be disabled or not.
+	/// \return boolean status of initialization. true if success; false
 	/// otherwise.
-	inline bool initialize(size_t order) {
-		if (order > max_order::value) {
-			return false;
-		}
-		list_ = nullptr;
-		order_ = order;
-		return true;
-	}
+	bool initialize(size_t order, bool disable_squashing = false);
 
 
 	/// \brief Get next free page block.
@@ -79,16 +69,7 @@ public:
 	/// The function allocates one page block from the allocator.
 	///
 	/// \return Allocated page block.
-	inline page_frame_t *get() {
-		item_type *free_item = list_;
-		if (free_item) {
-			list_ = list_->next;
-			free_item->next = nullptr;
-			//--stats.items;
-		}
-		sanity_check();
-		return reinterpret_cast<page_frame_t *>(free_item);
-	}
+	page_frame_t *get();
 
 
 	/// \brief Release page block.
@@ -99,155 +80,23 @@ public:
 	/// \param frame Pointer to the page block to be released.
 	/// \returns Pointer to squashed page block is squashing is possible;
 	/// nullptr otherwise.
-	page_frame_t *put(page_frame_t *frame) {
-		auto new_item = reinterpret_cast<item_type *>(frame);
-		new_item->next = nullptr;
-
-		if (list_ == nullptr) {
-			list_ = new_item;
-			//++stats.items;
-			sanity_check();
-			return nullptr;
-		}
-
-		item_type *last_lesser_item;
-		item_type *prelast_lesser_item;
-		find_last_lesser(new_item, last_lesser_item,
-			prelast_lesser_item);
-
-		if (order_ == max_order::value) {
-			// this FreeList is a free list of max order.
-			// it means that allocators of higher orders don't
-			// exist and we just puts frame to list.
-			new_item->next = last_lesser_item->next;
-			last_lesser_item->next = new_item;
-			++stats.items;
-			sanity_check();
-			return nullptr;
-		}
-
-
-		auto last_lesser_frame
-			= reinterpret_cast<page_frame_t *>(last_lesser_item);
-
-		if (last_lesser_item == nullptr) {
-			// Element should be placed to the beginning of the list.
-			if (are_consequent(frame, reinterpret_cast<page_frame_t *>(list_))
-				&& is_the_first_in_buddy(frame)) {
-				list_ = list_->next;
-				--stats.items;
-				sanity_check();
-				return frame;
-			} else {
-				new_item->next = list_;
-				list_ = new_item;
-				//++stats.items;
-				sanity_check();
-				return nullptr;
-			}
-		}
-
-		auto next_item = last_lesser_item->next;
-		auto next_frame = reinterpret_cast<page_frame_t *>(next_item);
-
-		if (are_consequent(last_lesser_frame, frame)
-				&& is_the_first_in_buddy(last_lesser_frame)) {
-			--stats.items;
-			if (prelast_lesser_item) {
-				prelast_lesser_item->next = last_lesser_item->next;
-				last_lesser_item->next = nullptr;
-				sanity_check();
-				return last_lesser_frame;
-			} else {
-				list_ = last_lesser_item->next;
-				last_lesser_item->next = nullptr;
-				sanity_check();
-				return last_lesser_frame;
-			}
-		}
-
-		if (are_consequent(frame, next_frame)
-				&& is_the_first_in_buddy(frame)) {
-			last_lesser_item->next = next_item->next;
-			--stats.items;
-			sanity_check();
-			return frame;
-		}
-
-
-		new_item->next = last_lesser_item->next;
-		last_lesser_item->next = new_item;
-		++stats.items;
-		sanity_check();
-
-		return nullptr;
-	}
+	page_frame_t *put(page_frame_t *frame);
 
 
 private:
 
+	/// Type of list element.
+	struct item_type;
+
+
 	/// Find last lesser.
-	void find_last_lesser(const item_type *new_item, item_type * &last_lesser,
-			item_type * &prelast_lesser) const {
-		prelast_lesser = nullptr;
-		last_lesser = nullptr;
-		for(auto item = list_; item && item < new_item;
-				item = item->next) {
-			prelast_lesser = last_lesser;
-			last_lesser = item;
-		}
-	}
+	void find_last_lesser(const item_type *new_item,
+			item_type * &last_lesser,
+			item_type * &prelast_lesser) const;
+
 
 	/// Run sanity check.
-	void sanity_check() const {
-		item_type *fast_iter = list_, *slow_iter = list_;
-		do {
-			if (slow_iter) {
-				slow_iter = slow_iter->next;
-			}
-			if (fast_iter) {
-				fast_iter = fast_iter->next;
-			}
-			if (fast_iter) {
-				fast_iter = fast_iter->next;
-			}
-		} while(slow_iter && fast_iter && slow_iter != fast_iter);
-
-		if (slow_iter == fast_iter && slow_iter) {
-			panic("FreeList loop detected!");
-		}
-	}
-
-
-	/// Check that frames are consequent.
-	bool are_consequent(const page_frame_t *first,
-			const page_frame_t *second)
-	{
-		const auto diff_frames = 1 << order_;
-		return second - first == diff_frames;
-	}
-
-
-	/// The function is the first in the pair of squashed pages.
-	bool is_the_first_in_buddy(const page_frame_t *frame) const {
-		auto numeric = reinterpret_cast<size_t>(frame);
-		auto divided = numeric / (PAGE_SIZE*(1 << order_));
-		return divided % 2 == 0;
-	}
-
-
-	/// Type of list element.
-	struct item_type {
-		/// Pointer to the next element.
-		item_type *next;
-
-
-		/// Clear internal allocator's data in item
-		/// for security reasons.
-		void clear() {
-			next = nullptr;
-		}
-	};
+	void sanity_check() const;
 
 
 	/// Pointer to the first item in the list.
@@ -257,20 +106,17 @@ private:
 	/// Order of the list.
 	size_t order_ = 0;
 
+	bool disable_squashing_ = false;
+
 
 	/// Output operator for \ref FreeLists
 	friend
-	cio::OutStream& operator<<(cio::OutStream& stream,
-				const FreeList<max_order::value>& fl) {
-		stream << "[FreeList<" << max_order::value << ">"
-			<< "(" << fl.order_ << "):";
-		for (auto *item = fl.list_; item; item = item->next) {
-			stream << " " << item;
-		}
-		stream << "]";
-		return stream;
-	}
+	cio::OutStream& memory::allocators::operator<<(cio::OutStream& stream,
+				const FreeList& fl);
 }; // class FreeList
+
+cio::OutStream& operator<<(cio::OutStream& stream,
+		const FreeList& fl);
 
 
 } // namespace allocators
